Rename robot implementation
